+++ src/Pages/StatisticsPage.tsx
@@ -2,7 +2,7 @@ import { ExerciseTitle } from '../Components/Shared/ExerciseTitle';
 import { ScrollableList } from '../Components/Shared/ScrollableList';
 import { HandInTimeStatistics } from '../Components/StatisticsPage/HandInTimeStatistics';
 import { PassedStatistics } from '../Components/StatisticsPage/PassedStatistics';
-import { createQueryListElements } from '../Util/createQueryListElements';
+import { createQueryListElements } from '../Util/createQueryStatisticsListElements';
 
 const STATISTICS_PAGE_BREADCRUMB = 'Statistics Page';
