+++ src/Util/createQueryStatisticsListElements.tsx
